Adding catch errors in script

+++ docker/blazingsql-build/build.sh
@@ -225,6 +225,7 @@ fi
 zeromq_install_dir=$workspace_dir/dependencies/zeromq_install_dir
 
 if [ ! -d $zeromq_install_dir ]; then
+    echo "### Zeromq - Star ###"
     cd $workspace_dir/dependencies/
     git clone https://github.com/zeromq/libzmq.git
     cd $workspace_dir/dependencies/libzmq
@@ -234,6 +235,7 @@ if [ ! -d $zeromq_install_dir ]; then
 
     mkdir -p $zeromq_build_dir
     cd $zeromq_build_dir
+    echo "### Zeromq - cmake ###"
     cmake -DCMAKE_BUILD_TYPE=Release \
           -DCMAKE_INSTALL_PREFIX:PATH=$zeromq_install_dir \
           -DCMAKE_C_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
@@ -241,7 +243,17 @@ if [ ! -d $zeromq_install_dir ]; then
           -DENABLE_CURVE=OFF \
           -DZMQ_BUILD_TESTS=OFF \
           ..  
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Zeromq - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Zeromq - end ###"
 fi
 
 # Package zeromq
@@ -256,6 +268,7 @@ cp -r $zeromq_install_dir/lib/* ${output}/zeromq/
 jzmq_install_dir=$workspace_dir/dependencies/jzmq_install_dir
 
 if [ ! -d $jzmq_install_dir ]; then
+    echo "### Jzmq - Start ###"
     rm -rf $jzmq_install_dir
     cd $workspace_dir/dependencies/
     git clone https://github.com/zeromq/jzmq.git
@@ -267,13 +280,19 @@ if [ ! -d $jzmq_install_dir ]; then
     LDFLAGS="-L$zeromq_install_dir/lib" CFLAGS="-I$zeromq_install_dir/include -D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" CXXFLAGS="-I$zeromq_install_dir/include -D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" ./configure --prefix=$jzmq_install_dir
     LDFLAGS="-L$zeromq_install_dir/lib" CFLAGS="-I$zeromq_install_dir/include -D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" CXXFLAGS="-I$zeromq_install_dir/include -D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" make -j4 install
     cd ..
+
+    echo "### Jzmq - mvn ###"
     mvn clean install -Dgpg.skip=true -DskipTests=true
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 fi
 
 # Package jzmq
 cd $workspace_dir
 mkdir -p ${output}/jzmq/
 cp -r $jzmq_install_dir/lib/* ${output}/jzmq/
+echo "### Jzmq - end ###"
 
 #END jzmq
 
@@ -282,6 +301,7 @@ cp -r $jzmq_install_dir/lib/* ${output}/jzmq/
 boost_install_dir=$workspace_dir/dependencies/boost_install_dir
 
 if [ ! -d $boost_install_dir ]; then
+    echo "### Boost - start ###"
     cd $workspace_dir/dependencies/
 
     boost_dir=$workspace_dir/dependencies/boost/
@@ -298,6 +318,10 @@ if [ ! -d $boost_install_dir ]; then
     cd $boost_build_dir
     ./bootstrap.sh --with-libraries=system,filesystem,regex,atomic,chrono,container,context,thread --with-icu --prefix=$boost_install_dir
     ./b2 install variant=release define=_GLIBCXX_USE_CXX11_ABI=0 stage cxxflags=-fPIC cflags=-fPIC link=static runtime-link=static threading=multi --exec-prefix=$boost_install_dir --prefix=$boost_install_dir -a
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+    echo "### Boost - end ###"
 fi
 
 #END boost
@@ -308,11 +332,16 @@ nvstrings_package=nvstrings-0.0.3-cuda9.2_py35_0
 nvstrings_install_dir=$workspace_dir/dependencies/$nvstrings_package
 
 if [ ! -d $nvstrings_install_dir ]; then
+    echo "### Nvstring - start ###"
     cd $workspace_dir/dependencies/
     nvstrings_url=https://anaconda.org/nvidia/nvstrings/0.0.3/download/linux-64/"$nvstrings_package".tar.bz2
     wget $nvstrings_url
     mkdir $nvstrings_package
     tar xvf "$nvstrings_package".tar.bz2 -C $nvstrings_package
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+    echo "### Nvstring - end ###"
 fi
 
 # Package nvstrings (always do this since this lib is needed by further deployment processes: conda, docker)
@@ -328,11 +357,16 @@ libhdfs3_package=libhdfs3
 libhdfs3_install_dir=$workspace_dir/dependencies/$libhdfs3_package
 
 if [ ! -d $libhdfs3_install_dir ]; then
+    echo "### Libhdfs3 - start ###"
     cd $workspace_dir/dependencies/
     libhdfs3_url=https://s3-us-west-2.amazonaws.com/blazing-public-downloads/_libs_/libhdfs3/libhdfs3.tar.gz
     wget $libhdfs3_url
     mkdir $libhdfs3_package
     tar xvf "$libhdfs3_package".tar.gz -C $libhdfs3_package
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+    echo "### Libhdfs3 - end ###"
 fi
 
 cd $workspace_dir
@@ -363,8 +397,15 @@ if [ ! -d $googletest_install_dir ]; then
           -DCMAKE_C_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
           -DCMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
           ..
-    echo "### Googletest - make ###"
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Googletest - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
     echo "### Googletest - End ###"
 fi
 
@@ -392,8 +433,16 @@ if [ ! -d $flatbuffers_install_dir ]; then
           -DCMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
           -DCMAKE_POSITION_INDEPENDENT_CODE=ON \
           ..
-    echo "### Flatbufferts - make ###"
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Flatbufferts - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     echo "### Flatbufferts - End ###"
 fi
 
@@ -413,9 +462,12 @@ if [ ! -d $lz4_install_dir ]; then
     lz4_build_dir=$workspace_dir/dependencies/lz4
 
     # NOTE build Boost with old C++ ABI _GLIBCXX_USE_CXX11_ABI=0 and with -fPIC
-    echo "### Lz4 - make ###"
+    echo "### Lz4 - make install ###"
     cd $lz4_build_dir
     CFLAGS="-D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC" CXXFLAGS="-D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC" PREFIX=$lz4_install_dir make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Lz4 - End ###"
 fi
@@ -446,8 +498,15 @@ if [ ! -d $zstd_install_dir ]; then
           -DCMAKE_POSITION_INDEPENDENT_CODE=ON \
           -DZSTD_BUILD_STATIC=ON \
           ..
-    echo "### Zstd - make ###"
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Zstd - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Zstd - End ###"
 fi
@@ -478,8 +537,15 @@ if [ ! -d $brotli_install_dir ]; then
           -DCMAKE_POSITION_INDEPENDENT_CODE=ON \
           -DBUILD_SHARED_LIBS=OFF \
           ..
-    echo "### Brotli - make ###"
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Brotli - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Brotli - End ###"
 fi
@@ -506,8 +572,11 @@ if [ ! -d $snappy_install_dir ]; then
     echo "### Snappy - Configure ###"
     CFLAGS="-D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" CXXFLAGS="-D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" ./configure --prefix=$snappy_install_dir
 
-    echo "### Snappy - make ###"
+    echo "### Snappy - make install ###"
     CFLAGS="-D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" CXXFLAGS="-D_GLIBCXX_USE_CXX11_ABI=0 -O3 -fPIC -O2" make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Snappy - End ###"
 fi
@@ -551,9 +620,15 @@ if [ ! -d $thrift_install_dir ]; then
           -DCMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
           -DBOOST_ROOT=$boost_install_dir \
           ..
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
-    echo "### Thrift - make ###"
+    echo "### Thrift - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Thrift - end ###"
 fi
@@ -613,8 +688,15 @@ if [ ! -d $arrow_install_dir ]; then
         -DARROW_TENSORFLOW=ON \
         -DARROW_PARQUET=ON \
         ..
-    echo "### Arrow - make ###"
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Arrow - make install ###"
     make -j4 install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Arrow - end ###"
 fi
@@ -648,8 +730,15 @@ if [ ! -d $aws_sdk_cpp_build_dir ]; then
         -DCMAKE_C_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
         -DCMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
         ..
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     echo "### Aws sdk - make ###"
     make -j4
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Arrow - end ###"
 fi
@@ -698,8 +787,11 @@ if [ $cudf_enable == true ]; then
         -DCMAKE_BUILD_TYPE=Release  \
         -DCMAKE_INSTALL_PREFIX:PATH=$libgdf_install_dir  \
         ..
-    echo "### CUDF - make ###"
+    echo "### CUDF - make install ###"
     make -j$cudf_parallel install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
     
     #TODO remove this patch once cudf can install rmm
     cp $cudf_current_dir/cudf/$libgdf_dir/src/rmm/memory.h $libgdf_install_dir/include
@@ -765,11 +857,23 @@ if [ $blazingdb_protocol_enable == true ]; then
           -DCMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
 	  -DZEROMQ_INSTALL_DIR=$zeromq_install_dir \
           ..
-    echo "### Protocol - make ###"
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
+    echo "### Protocol - make install ###"
     make -j$blazingdb_protocol_parallel install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
     
     cd $blazingdb_protocol_current_dir/blazingdb-protocol/java
+    echo "### Protocol - mvn ###"
     mvn clean install -Dmaven.test.skip=true
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     blazingdb_protocol_java_build_dir=$blazingdb_protocol_current_dir/blazingdb-protocol/java/target/
     
     #END blazingdb-protocol
@@ -827,8 +931,15 @@ if [ $blazingdb_io_enable == true ]; then
           -DCMAKE_C_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
           -DCMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0 \
           ..
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     echo "### Blazingdb IO - make ###"
     make -j$blazingdb_io_parallel install
+    if [ $? != 0 ]; then
+      exit 1
+    fi
     
     echo "### Blazingdb IO - end ###"
     #END blazingdb-io
@@ -899,9 +1010,15 @@ if [ $blazingdb_ral_enable == true ]; then
           -DCUDA_DEFINES=$blazingdb_ral_definitions \
           -DCXX_DEFINES=$blazingdb_ral_definitions \
           ..
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Ral - make ###"
     make -j$blazingdb_ral_parallel
+    if [ $? != 0 ]; then
+      exit 1
+    fi
     
     #END blazingdb-ral
     
@@ -909,6 +1026,9 @@ if [ $blazingdb_ral_enable == true ]; then
     cd $workspace_dir
     blazingdb_ral_artifact_name=testing-libgdf
     cp $blazingdb_ral_build_dir/$blazingdb_ral_artifact_name $output
+    if [ $? != 0 ]; then
+      exit 1
+    fi
 
     echo "### Ral - end ###"
 fi
@@ -957,8 +1077,15 @@ if [ $blazingdb_orchestrator_enable == true ]; then
     cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE  \
           -DBLAZINGDB_PROTOCOL_BRANCH=$blazingdb_protocol_branch \
           ..
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     echo "### Orchestrator - make ###"
     make -j$blazingdb_orchestrator_parallel
+    if [ $? != 0 ]; then
+      exit 1
+    fi
     
     #END blazingdb-orchestrator
     
@@ -1000,6 +1127,10 @@ if [ $blazingdb_calcite_enable == true ]; then
     
     echo "### Calcite - mvn clean install ###"
     mvn clean install -Dmaven.test.skip=true
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     blazingdb_calcite_build_dir=$blazingdb_calcite_current_dir/blazingdb-calcite/blazingdb-calcite-application/target/
     
     #END blazingdb-calcite
@@ -1046,6 +1177,10 @@ if [ $pyblazing_enable == true ]; then
     cd $workspace_dir
     mkdir -p ${output}/pyBlazing/
     cp -r $pyblazing_current_dir/pyBlazing/* ${output}/pyBlazing/
+    if [ $? != 0 ]; then
+      exit 1
+    fi
+
     rm -rf ${output}/pyBlazing/.git/
     echo "### Pyblazing - end ###"
 fi
